Simplified pre/post command implementation.
This implementation runs the pre/post command immediately before and after
making a snapshot, rather than at the beginning/end of a backup. In addition,
it will run once per dataset and pass the name of the dataset as the first
(and only) argument to the command, allowing scripts to trigger only on some
datasets.

Unlike the previous implementation. This implementation is entirely based in
the agent and requires no specific support from the zetaback server.