Exclude deleted filesystems from violators list
Previously, if you removed a filesystem from a server being backed up, it
would eventually show up in the violators list, because it wasn't being backed
up (due to no longer existing). This meant that a filesystem showing up in the
violators list did not necessarily signify a problem with backups, and in fact
can be a common occurrence on systems where the filesystem list changes often
(e.g. OS upgrades, removing old zones).

Now, every time a backup is taken, the list of filesystems that exist on the
server is cached, and any filesystems not in the list when viewing violators
are not included.

The old behavior is still available with the -svv option.