Add a setting to limit the number of zetaback processes running
If a backup takes longer than the interval between zetaback runs as specified
in the crontab, then multiple zetaback instances will begin running. This
isn't a bad thing, but can become problematic if too many start. This setting
will limit the number of running zetaback processes to the amount specified.
This makes use of numbered lock files - zetaback will try each on in turn
until it can lock one of the files, or give up if it tries them all.