Allow restoring multiple filesystems in bulk
If you specify ranges when choosing the backups, it will restore them in turn.
Rather than force you to pick times for each backup, it will automatically
choose the latest backup for each filesytem in this mode.
Note that you need to restore parent filesystems first before child
filesystems. If you pick ranges that are always in numerical order, this will
be the case.