Add option to exclude alternate/inactive boot environments from a backup
This option obtains the ID of the currently active boot environment, and then
looks for filesystems eith either the BE uuid or parent BE uuid (for zone
filesystems) zfs properties set. If they are set, but don't match the uuid of
the current boot environment, they are excluded. This option is configurable
in zetaback_agent.conf.